backend/__tests__/tracesController.test.ts
import request from 'supertest'; // For making mock HTTP requests
import { app } from '../server'; // Assuming you have an Express app instance
import * as zlib from 'zlib'; // For mocking gzip decompression
import fs from 'fs'; // For mocking filesystem operations
import * as protobuf from 'protobufjs'; // For mocking protobuf decoding

jest.mock('fs');
jest.mock('zlib');
jest.mock('protobufjs');

describe('tracesController', () => {
  describe('decompressRequest', () => {
    it('should decompress gzip encoded request body', async () => {
      const mockGzipContent = Buffer.from('mock compressed data');
      zlib.gunzip.mockImplementation((buf, callback) => {
        callback(null, Buffer.from('mock decompressed data'));
      });

      await request(app)
        .post('/traces')
        .send(mockGzipContent)
        .set('Content-Encoding', 'gzip')
        .expect(200);

      expect(zlib.gunzip).toHaveBeenCalled();
      // Further assertions to verify the request body was correctly decompressed can be added here
    });
  });

  describe('decodeTraceData', () => {
    it('should decode protobuf encoded trace data', async () => {
      protobuf.load.mockResolvedValue({
        lookupType: () => ({
          decode: jest.fn().mockReturnValue({}),
          toObject: jest.fn().mockReturnValue({ decodedData: 'mock decoded data' })
        })
      });

      // Mock the request to simulate sending protobuf encoded data
      const mockRequest = { /* Simulate protobuf encoded data */ };

      await request(app)
        .post('/traces')
        .send(mockRequest)
        .expect(200);

      // Verify that protobuf decoding was attempted
      expect(protobuf.load).toHaveBeenCalled();
      // Additional assertions to verify decoding can be added here
    });
  });

  describe('storeTraces', () => {
    it('should store decoded trace data', async () => {
      fs.readFileSync.mockReturnValue(JSON.stringify([])); // Mock existing data as empty array
      fs.writeFileSync.mockImplementation(() => {});

      const mockDecodedData = { resourceSpans: [{ scopeSpans: [{ spans: ['span1', 'span2'] }] }] };

      await request(app)
        .post('/traces')
        .send({ decodedData: mockDecodedData }) // This assumes the body parser middleware correctly sets req.body
        .expect(200);

      expect(fs.writeFileSync).toHaveBeenCalled();
      // Further assertions to verify the data was correctly stored can be added here
    });
  });

  // Additional tests for printTraces and any error handling can be added here
});
